+++ src/base/debris.js
@@ -16,27 +16,32 @@ class Debris extends Base1 {
     this.kind = props.kind;
     this.vertices = props.vertices;
     this.e = 100;
+    this.initialE = 100;
     this.decay = 0.01;
+    if (this.kind === kinds.kAsteroidDebris) {
+      this.e = 200;
+      this.initialE = 200;
+    }
     this.width = props.width;
     this.color = props.color;
   }
 
   static getMeshes(kind, vertices, width) {
     let meshes = [];
-    if (kind === kinds.kShipDebris) {
-      for (let i = 1; i < vertices.length; i++) {
-        const v0 = vertices[i - 1];
-        const v1 = vertices[i];
-        const mesh = new Mesh({
-          kind: Meshes.kLine,
-          vertices: [v0, v1],
-          color: 0xffffff,
-          width: width ?? 10,
-        });
+    //if (kind === kinds.kShipDebris) {
+    for (let i = 1; i < vertices.length; i++) {
+      const v0 = vertices[i - 1];
+      const v1 = vertices[i];
+      const mesh = new Mesh({
+        kind: Meshes.kLine,
+        vertices: [v0, v1],
+        color: 0xffffff,
+        width: width ?? 10,
+      });
 
-        meshes.push(mesh);
-      }
+      meshes.push(mesh);
     }
+    //}
     return meshes;
   }
 
@@ -61,8 +66,14 @@ class Debris extends Base1 {
       p.shiftX = 0;
       p.shiftY = 0;
       p.shiftR = 0;
-      p.rSpeed = 0.01;
-      p.shiftSpeed = 0.1 + 0.5 * Math.random();
+      if (this.kind === kinds.kAsteroidDebris) {
+        p.rSpeed = 0.02;
+        p.shiftSpeed = 0.9 + 0.9 * Math.random();
+      } else {
+        p.rSpeed = 0.01;
+        p.shiftSpeed = 0.1 + 0.5 * Math.random();
+      }
+
       p.shiftSpeedDecay = 0.001;
       p.rSpeedDecay = 0.00005;
       this.presentations.push(p);
@@ -102,10 +113,12 @@ class Debris extends Base1 {
     }
     super.move(t); // This will move globally
 
-    this.e -= 0.5;
-    this.e = Math.max(1, this.e);
-    const ne = Math.max(0, Math.min(1, this.e / 100));
-    const g = 100 + 155 * ne;
+    this.e -= 0.1;
+    if (this.kind === kinds.kShipDebris) {
+      this.e = Math.max(1, this.e);
+    }
+    const ne = Math.max(0, Math.min(1, this.e / this.initialE));
+    const g = (this.kind === kinds.kShipDebris ? 100 : 0) + 155 * ne;
     const hexColor = (g << 16) | (g << 8) | g;
     for (let presentation of this.presentations) {
       if (!presentation) {