+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.8.3
+
+- Remove auto-repair
+- Add a proper state machine (might be _buggy_)
+
 # 0.8.2
 
 - Ship debris